Introduce BidiTransport based on BiDi messages
