+++ src/bidiMapper/domains/log/logHelper.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {CommonDataTypes} from '../protocol/bidiProtocolTypes';
+import type {CommonDataTypes} from '../../../protocol/types';
 
 const specifiers = ['%s', '%d', '%i', '%f', '%o', '%O', '%c'];
