+++ src/protocol/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../configs/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "../../lib/cjs/protocol"
+  },
+  "references": [{"path": "../utils/tsconfig.json"}]
+}
